'curry' the xml
Now that we only ever use it as a parameter, we can actually remove it
as a parameter and put it directly into our xml methods.

This is a reference to curried functions in Haskell:

add :: Integer -> Integer -> Integer
add x y = x + y

add_five = add 5

add_five is a 'curried' version of foo. You can do this in Ruby too,
but it's so much more elegant in Haskell.