Simple instance variable caching.
The problem with the last commit is that we ended up making a LOT MORE
requests than we should have, due to making excessive XML objects. Then
inside of the XML object, we referred to the virtual attribute that
makes a Request multiple times. Ack! Now we don't do that any more.

Note that HTTP caching would be the 'right' thing to do here. I don't
want to get into that yet, so we're just using instance variables.